Adding master-slave merge capabilities to remote repos. We do this by merging in a non-bare repo.
Writing an actual 3-way merge algorithm in dulwich is difficult.
Additionally, bare repositories have no index or working tree, so
shelling out doesn't work. To work around this, we currently create a
cloned slave of the remote repository on the same filesystem,
fetch the new change, attempt to merge it to the requested ref to merge
into, and push back into the repository. This is done by shelling out to
the command line.